Move ha_discovery property changed handling one level up
It was located in the add_dbus_to_mqtt_mapping function that was called
per changed item. It is now located in the add_dbus_to_mqtt_mapping that
processes all changed items. This allows for some simplification and
cleanup.

One example is that is calls dbus_service_items[service]. This causes
the entry to always exists in dbus_service_items. So a find on service
in dbus_service_items can be removed. This leads to the remove of the
publishSensorEntity function.

This also paves the way to change the handling to a device based
handling instead of an entity based handling.